refactor(community): move indexer poll helper out of the dialog component
Fast Refresh requires component files to export only components. The
waitForCommunityIndexed helper and its poll constants were exported from
CommunityDialog.tsx, tripping React Doctor's only-export-components rule
(the quality-gate regression). Move them to utilities/waitForCommunityIndexed.ts
with its test alongside; the dialog now imports the helper. No behavior change.
